+++ library/src/main/java/com/bkromhout/rrvl/RealmSimpleItemTouchHelperCallback.java
@@ -4,7 +4,7 @@ import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
 
 /**
- * Implementation of {@link ItemTouchHelper.Callback} for supporting drag and drop. Adapted from <a
+ * Implementation of {@code ItemTouchHelper.Callback} for supporting drag and drop. Adapted from <a
  * href="https://medium.com/@ipaulpro/drag-and-swipe-with-recyclerview-b9456d2b1aaf">this article</a>.
  */
 public class RealmSimpleItemTouchHelperCallback extends ItemTouchHelper.Callback {
@@ -45,7 +45,7 @@ public class RealmSimpleItemTouchHelperCallback extends ItemTouchHelper.Callback
 
     @Override
     public int getMovementFlags(RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder) {
-        return makeMovementFlags(dragAndDrop ? ItemTouchHelper.UP | ItemTouchHelper.DOWN : 0, 0);
+        return makeMovementFlags(dragAndDrop ? (ItemTouchHelper.UP | ItemTouchHelper.DOWN) : 0, 0);
     }
 
     @Override